+++ packages/brand-components/src/Header/HeaderTab/index.tsx
@@ -1,0 +1,26 @@
+import React, { HTMLAttributes } from 'react';
+import cx from 'classnames';
+import s from './styles.module.scss';
+
+export type HeaderTabProps = HTMLAttributes<HTMLElement> & {
+  className?: string;
+  id?: string;
+  testId?: string;
+};
+
+export const HeaderTab: React.FC<HeaderTabProps> = ({
+  className,
+  children,
+  id,
+  testId,
+}) => {
+  const classes = cx(s.headerTab, className);
+
+  return (
+    <div id={id} className={classes} data-testid={testId}>
+      {children}
+    </div>
+  );
+};
+
+export default HeaderTab;