+++ packages/gamut/src/Icon/icons/GitIcon.tsx
@@ -2,7 +2,7 @@ import React, { SVGProps } from 'react';
 
 import { defaultIconProps } from '../defaultIconProps';
 
-export const GitIcon: React.FC<SVGProps<SVGSVGElement>> = props => {
+export const GitIcon: React.FC<SVGProps<SVGSVGElement>> = (props) => {
   return (
     <svg {...props}>
       <title>Git Icon</title>