+++ packages/gamut/src/Icon/icons/PlayIcon.tsx
@@ -2,7 +2,7 @@ import React, { SVGProps } from 'react';
 
 import { defaultIconProps } from '../defaultIconProps';
 
-export const PlayIcon: React.FC<SVGProps<SVGSVGElement>> = props => {
+export const PlayIcon: React.FC<SVGProps<SVGSVGElement>> = (props) => {
   return (
     <svg {...props}>
       <title>Play Icon</title>