+++ packages/webpack-config/src/config/css-extracted.js
@@ -1,11 +1,13 @@
 const MiniCssExtractPlugin = require('mini-css-extract-plugin');
 const merge = require('webpack-merge');
 const loaders = require('../loaders');
+const ENV = require('../lib/env');
+const PROD = ENV === 'production';
 
-const cssExtracted = options => {
+const cssExtracted = (options) => {
   const defaultExtractOptions = {
-    filename: '[name].css',
-    esModule: true,
+    filename: PROD ? '[name].[contenthash].css' : '[name].css',
+    chunkFilename: PROD ? '[id].[contenthash].chunk.css' : '[id].chunk.css',
     ignoreOrder: true,
   };